requirements.txt
mcp
chromadb

PyExifTool
